Move LocalServerPort to web-context. Fixes gh-18392
