+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/web/context/LocalServerPort.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.web.server;
+package org.springframework.boot.web.context;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
